+++ android/src/main/java/com/hoxfon/react/RNTwilioVoice/TwilioVoiceModule.java
@@ -43,9 +43,12 @@ import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.InstanceIdResult;
+import com.twilio.voice.AcceptOptions;
 import com.twilio.voice.Call;
 import com.twilio.voice.CallException;
 import com.twilio.voice.CallInvite;
+import com.twilio.voice.CancelledCallInvite;
+import com.twilio.voice.ConnectOptions;
 import com.twilio.voice.LogLevel;
 import com.twilio.voice.RegistrationException;
 import com.twilio.voice.RegistrationListener;
@@ -59,6 +62,8 @@ import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_CONNECTION_DID_D
 import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_DID_RECEIVE_INCOMING;
 import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_NOT_READY;
 import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_DEVICE_READY;
+import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_CALL_STATE_RINGING;
+import static com.hoxfon.react.RNTwilioVoice.EventManager.EVENT_CALL_INVITE_CANCELLED;
 
 public class TwilioVoiceModule extends ReactContextBaseJavaModule implements ActivityEventListener, LifecycleEventListener {
 
@@ -78,6 +83,8 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
     public static final String INCOMING_CALL_INVITE          = "INCOMING_CALL_INVITE";
     public static final String INCOMING_CALL_NOTIFICATION_ID = "INCOMING_CALL_NOTIFICATION_ID";
     public static final String NOTIFICATION_TYPE             = "NOTIFICATION_TYPE";
+    public static final String CANCELLED_CALL_INVITE         = "CANCELLED_CALL_INVITE";
+
 
     public static final String ACTION_INCOMING_CALL = "com.hoxfon.react.TwilioVoice.INCOMING_CALL";
     public static final String ACTION_FCM_TOKEN     = "com.hoxfon.react.TwilioVoice.ACTION_FCM_TOKEN";
@@ -85,6 +92,7 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
     public static final String ACTION_ANSWER_CALL   = "com.hoxfon.react.TwilioVoice.ANSWER_CALL";
     public static final String ACTION_REJECT_CALL   = "com.hoxfon.react.TwilioVoice.REJECT_CALL";
     public static final String ACTION_HANGUP_CALL   = "com.hoxfon.react.TwilioVoice.HANGUP_CALL";
+    public static final String ACTION_CANCEL_CALL_INVITE = "com.hoxfon.react.TwilioVoice.CANCEL_CALL_INVITE";
     public static final String ACTION_CLEAR_MISSED_CALLS_COUNT = "com.hoxfon.react.TwilioVoice.CLEAR_MISSED_CALLS_COUNT";
 
     public static final String CALL_SID_KEY = "CALL_SID";
@@ -217,6 +225,34 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
 
     private Call.Listener callListener() {
         return new Call.Listener() {
+            /*
+             * This callback is emitted once before the Call.Listener.onConnected() callback when
+             * the callee is being alerted of a Call. The behavior of this callback is determined by
+             * the answerOnBridge flag provided in the Dial verb of your TwiML application
+             * associated with this client. If the answerOnBridge flag is false, which is the
+             * default, the Call.Listener.onConnected() callback will be emitted immediately after
+             * Call.Listener.onRinging(). If the answerOnBridge flag is true, this will cause the
+             * call to emit the onConnected callback only after the call is answered.
+             * See answeronbridge for more details on how to use it with the Dial TwiML verb. If the
+             * twiML response contains a Say verb, then the call will emit the
+             * Call.Listener.onConnected callback immediately after Call.Listener.onRinging() is
+             * raised, irrespective of the value of answerOnBridge being set to true or false
+             */
+            @Override
+            public void onRinging(Call call) {
+                // TODO test this with JS app
+                if (BuildConfig.DEBUG) {
+                    Log.d(TAG, "CALL RINGING callListener().onRinging call state = "+call.getState());
+                    Log.d(TAG, call.toString());
+                }
+                WritableMap params = Arguments.createMap();
+                if (call != null) {
+                    params.putString("call_sid",   call.getSid());
+                    params.putString("call_from",  call.getFrom());
+                    params.putString("call_state", call.getState().name());
+                }
+                eventManager.sendEvent(EVENT_CALL_STATE_RINGING, params);
+            }
             @Override
             public void onConnected(Call call) {
                 if (BuildConfig.DEBUG) {
@@ -244,6 +280,27 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                 }
                 eventManager.sendEvent(EVENT_CONNECTION_DID_CONNECT, params);
             }
+
+//            Prepare methods Twilio for v4
+//            @Override
+//            public void onReconnecting(Call call, CallException callException) {
+//                if (BuildConfig.DEBUG) {
+//                    Log.d(TAG, "CALL RECONNECTING callListener().onReconnecting call state = "+call.getState());
+//                }
+//                // TODO implement
+////                eventManager.sendEvent(XXX, params);
+//
+//            }
+//
+//            @Override
+//            public void onReconnected(Call call) {
+//                if (BuildConfig.DEBUG) {
+//                    Log.d(TAG, "CALL RECONNECTED callListener().onReconnected call state = "+call.getState());
+//                }
+//
+//                // TODO implement
+////                eventManager.sendEvent(XXX, params);
+//            }
 
             @Override
             public void onDisconnected(Call call, CallException error) {
@@ -319,6 +376,7 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         if (!isReceiverRegistered) {
             IntentFilter intentFilter = new IntentFilter();
             intentFilter.addAction(ACTION_INCOMING_CALL);
+            intentFilter.addAction(ACTION_CANCEL_CALL_INVITE);
             intentFilter.addAction(ACTION_MISSED_CALL);
             LocalBroadcastManager.getInstance(getReactApplicationContext()).registerReceiver(
                     voiceBroadcastReceiver, intentFilter);
@@ -381,19 +439,13 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
     }
 
     private void handleIncomingCallIntent(Intent intent) {
-        if (intent == null || intent.getAction() == null) {
-            Log.e(TAG, "handleIncomingCallIntent intent is null");
-            return;
-        }
-
         if (intent.getAction().equals(ACTION_INCOMING_CALL)) {
+            if (BuildConfig.DEBUG) {
+                Log.d(TAG, "handleIncomingCallIntent");
+            }
             activeCallInvite = intent.getParcelableExtra(INCOMING_CALL_INVITE);
-
-            if (activeCallInvite != null && (activeCallInvite.getState() == CallInvite.State.PENDING)) {
+            if (activeCallInvite != null) {
                 callAccepted = false;
-                if (BuildConfig.DEBUG) {
-                    Log.d(TAG, "handleIncomingCallIntent state = PENDING");
-                }
                 SoundPoolManager.getInstance(getReactApplicationContext()).playRinging();
 
                 if (getReactApplicationContext().getCurrentActivity() != null) {
@@ -411,58 +463,36 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                     WritableMap params = Arguments.createMap();
                     params.putString("call_sid", activeCallInvite.getCallSid());
                     params.putString("call_from", activeCallInvite.getFrom());
-                    params.putString("call_to", activeCallInvite.getTo());
-                    params.putString("call_state", activeCallInvite.getState().name());
+                    params.putString("call_to", activeCallInvite.getTo()); // this is not needed
+                    // TODO check if this is needed
+                    // params.putString("call_state", "PENDING");
                     eventManager.sendEvent(EVENT_DEVICE_DID_RECEIVE_INCOMING, params);
                 }
-
-
             } else {
+                // TODO evaluate what more is needed at this point?
+                Log.e(TAG, "ACTION_INCOMING_CALL but not active call");
+            }
+        } else if (intent.getAction().equals(ACTION_CANCEL_CALL_INVITE)) {
+            SoundPoolManager.getInstance(getReactApplicationContext()).stopRinging();
+            if (BuildConfig.DEBUG) {
+                Log.d(TAG, "activeCallInvite was cancelled by " + activeCallInvite.getFrom());
+            }
+            if (!callAccepted) {
                 if (BuildConfig.DEBUG) {
-                    Log.d(TAG, "====> BEGIN handleIncomingCallIntent when activeCallInvite != PENDING");
+                    Log.d(TAG, "creating a missed call");
                 }
-                // this block is executed when the callInvite is cancelled and:
-                //   - the call is answered (activeCall != null)
-                //   - the call is rejected
-
-                SoundPoolManager.getInstance(getReactApplicationContext()).stopRinging();
-
-                // the call is not active yet
-                if (activeCall == null) {
-
-                    if (activeCallInvite != null) {
-                        if (BuildConfig.DEBUG) {
-                            Log.d(TAG, "activeCallInvite state = " + activeCallInvite.getState());
-                        }
-                        if (BuildConfig.DEBUG) {
-                            Log.d(TAG, "activeCallInvite was cancelled by " + activeCallInvite.getFrom());
-                        }
-                        if (!callAccepted) {
-                            if (BuildConfig.DEBUG) {
-                                Log.d(TAG, "creating a missed call, activeCallInvite state: " + activeCallInvite.getState());
-                            }
-                            callNotificationManager.createMissedCallNotification(getReactApplicationContext(), activeCallInvite);
-                            int appImportance = callNotificationManager.getApplicationImportance(getReactApplicationContext());
-                            if (appImportance != RunningAppProcessInfo.IMPORTANCE_BACKGROUND) {
-                                WritableMap params = Arguments.createMap();
-                                params.putString("call_sid", activeCallInvite.getCallSid());
-                                params.putString("call_from", activeCallInvite.getFrom());
-                                params.putString("call_to", activeCallInvite.getTo());
-                                params.putString("call_state", activeCallInvite.getState().name());
-                                eventManager.sendEvent(EVENT_CONNECTION_DID_DISCONNECT, params);
-                            }
-                        }
-                    }
-                    clearIncomingNotification(activeCallInvite);
-                } else {
-                    if (BuildConfig.DEBUG) {
-                        Log.d(TAG, "activeCallInvite was answered. Call " + activeCall);
-                    }
-                }
-                if (BuildConfig.DEBUG) {
-                    Log.d(TAG, "====> END");
+                callNotificationManager.createMissedCallNotification(getReactApplicationContext(), activeCallInvite);
+                int appImportance = callNotificationManager.getApplicationImportance(getReactApplicationContext());
+                if (appImportance != RunningAppProcessInfo.IMPORTANCE_BACKGROUND) {
+                    WritableMap params = Arguments.createMap();
+                    params.putString("call_sid", activeCallInvite.getCallSid());
+                    params.putString("call_from", activeCallInvite.getFrom());
+                    params.putString("call_to", activeCallInvite.getTo());
+                    params.putString("call_state", Call.State.DISCONNECTED.toString());
+                    eventManager.sendEvent(EVENT_CONNECTION_DID_DISCONNECT, params);
                 }
             }
+            clearIncomingNotification(activeCallInvite.getCallSid());
         } else if (intent.getAction().equals(ACTION_FCM_TOKEN)) {
             if (BuildConfig.DEBUG) {
                 Log.d(TAG, "handleIncomingCallIntent ACTION_FCM_TOKEN");
@@ -476,11 +506,21 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         @Override
         public void onReceive(Context context, Intent intent) {
             String action = intent.getAction();
+            if (BuildConfig.DEBUG) {
+                Log.d(TAG, "VoiceBroadcastReceiver.onReceive "+action+". Intent "+ intent.getExtras());
+            }
             if (action.equals(ACTION_INCOMING_CALL)) {
-                if (BuildConfig.DEBUG) {
-                    Log.d(TAG, "VoiceBroadcastReceiver.onReceive ACTION_INCOMING_CALL. Intent "+ intent.getExtras());
-                }
                 handleIncomingCallIntent(intent);
+            } else if (action.equals(ACTION_CANCEL_CALL_INVITE)) {
+                CancelledCallInvite cancelledCallInvite = intent.getParcelableExtra(CANCELLED_CALL_INVITE);
+                clearIncomingNotification(cancelledCallInvite.getCallSid());
+                WritableMap params = Arguments.createMap();
+                if (cancelledCallInvite != null) {
+                    params.putString("call_sid", cancelledCallInvite.getCallSid());
+                    params.putString("call_from", cancelledCallInvite.getFrom());
+                    params.putString("call_to", cancelledCallInvite.getTo());
+                }
+                eventManager.sendEvent(EVENT_CALL_INVITE_CANCELLED, params);
             } else if (action.equals(ACTION_MISSED_CALL)) {
                 SharedPreferences sharedPref = getReactApplicationContext().getSharedPreferences(PREFERENCE_KEY, Context.MODE_PRIVATE);
                 SharedPreferences.Editor sharedPrefEditor = sharedPref.edit();
@@ -514,21 +554,19 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         promise.resolve(params);
     }
 
-    private void clearIncomingNotification(CallInvite callInvite) {
+    private void clearIncomingNotification(String callSid) {
         if (BuildConfig.DEBUG) {
-            Log.d(TAG, "clearIncomingNotification() callInvite state: "+ callInvite.getState());
+            Log.d(TAG, "clearIncomingNotification() callSid: "+ callSid);
         }
-        if (callInvite != null && callInvite.getCallSid() != null) {
-            // remove incoming call notification
-            String notificationKey = INCOMING_NOTIFICATION_PREFIX + callInvite.getCallSid();
-            int notificationId = 0;
-            if (TwilioVoiceModule.callNotificationMap.containsKey(notificationKey)) {
-                notificationId = TwilioVoiceModule.callNotificationMap.get(notificationKey);
-            }
-            callNotificationManager.removeIncomingCallNotification(getReactApplicationContext(), null, notificationId);
-            TwilioVoiceModule.callNotificationMap.remove(notificationKey);
+        // remove incoming call notification
+        String notificationKey = INCOMING_NOTIFICATION_PREFIX + callSid;
+        int notificationId = 0;
+        if (TwilioVoiceModule.callNotificationMap.containsKey(notificationKey)) {
+            notificationId = TwilioVoiceModule.callNotificationMap.get(notificationKey);
         }
-//        activeCallInvite = null;
+        callNotificationManager.removeIncomingCallNotification(getReactApplicationContext(), null, notificationId);
+        TwilioVoiceModule.callNotificationMap.remove(notificationKey);
+        activeCallInvite = null;
     }
 
     /*
@@ -554,38 +592,36 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                             if (BuildConfig.DEBUG) {
                                 Log.d(TAG, "Registering with FCM");
                             }
-                            Voice.register(getReactApplicationContext(), accessToken, Voice.RegistrationChannel.FCM, fcmToken, registrationListener);
+                            Voice.register(accessToken, Voice.RegistrationChannel.FCM, fcmToken, registrationListener);
                         }
                     }
                 });
-
     }
 
     @ReactMethod
     public void accept() {
         callAccepted = true;
         SoundPoolManager.getInstance(getReactApplicationContext()).stopRinging();
-        if (activeCallInvite != null){
-            if (activeCallInvite.getState() == CallInvite.State.PENDING) {
-                if (BuildConfig.DEBUG) {
-                    Log.d(TAG, "accept() activeCallInvite.getState() PENDING");
-                }
-                activeCallInvite.accept(getReactApplicationContext(), callListener);
-                clearIncomingNotification(activeCallInvite);
-            } else {
-                // when the user answers a call from a notification before the react-native App
-                // is completely initialised, and the first event has been skipped
-                // re-send connectionDidConnect message to JS
-                WritableMap params = Arguments.createMap();
-                params.putString("call_sid",   activeCallInvite.getCallSid());
-                params.putString("call_from",  activeCallInvite.getFrom());
-                params.putString("call_to",    activeCallInvite.getTo());
-                params.putString("call_state", activeCallInvite.getState().name());
-                callNotificationManager.createHangupLocalNotification(getReactApplicationContext(),
-                        activeCallInvite.getCallSid(),
-                        activeCallInvite.getFrom());
-                eventManager.sendEvent(EVENT_CONNECTION_DID_CONNECT, params);
+        if (activeCallInvite != null) {
+            if (BuildConfig.DEBUG) {
+                Log.d(TAG, "accept()");
             }
+            AcceptOptions acceptOptions = new AcceptOptions.Builder().build();
+            activeCallInvite.accept(getReactApplicationContext(), acceptOptions, callListener);
+            clearIncomingNotification(activeCallInvite.getCallSid());
+
+            // TODO check whether this block is needed
+//            // when the user answers a call from a notification before the react-native App
+//            // is completely initialised, and the first event has been skipped
+//            // re-send connectionDidConnect message to JS
+//            WritableMap params = Arguments.createMap();
+//            params.putString("call_sid",   activeCallInvite.getCallSid());
+//            params.putString("call_from",  activeCallInvite.getFrom());
+//            params.putString("call_to",    activeCallInvite.getTo());
+//            callNotificationManager.createHangupLocalNotification(getReactApplicationContext(),
+//                    activeCallInvite.getCallSid(),
+//                    activeCallInvite.getFrom());
+//            eventManager.sendEvent(EVENT_CONNECTION_DID_CONNECT, params);
         } else {
             eventManager.sendEvent(EVENT_CONNECTION_DID_DISCONNECT, null);
         }
@@ -596,13 +632,15 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         callAccepted = false;
         SoundPoolManager.getInstance(getReactApplicationContext()).stopRinging();
         WritableMap params = Arguments.createMap();
-        if (activeCallInvite != null){
+        if (activeCallInvite != null) {
             params.putString("call_sid",   activeCallInvite.getCallSid());
             params.putString("call_from",  activeCallInvite.getFrom());
             params.putString("call_to",    activeCallInvite.getTo());
-            params.putString("call_state", activeCallInvite.getState().name());
+            params.putString("call_state", "DISCONNECTED");
+            // TODO check if DISCONNECTED should be REJECTED
+            // params.putString("call_state", "REJECTED");
             activeCallInvite.reject(getReactApplicationContext());
-            clearIncomingNotification(activeCallInvite);
+            clearIncomingNotification(activeCallInvite.getCallSid());
         }
         eventManager.sendEvent(EVENT_CONNECTION_DID_DISCONNECT, params);
     }
@@ -612,12 +650,12 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         callAccepted = false;
         SoundPoolManager.getInstance(getReactApplicationContext()).stopRinging();
         WritableMap params = Arguments.createMap();
-        if (activeCallInvite != null){
+        if (activeCallInvite != null) {
             params.putString("call_sid",   activeCallInvite.getCallSid());
             params.putString("call_from",  activeCallInvite.getFrom());
             params.putString("call_to",    activeCallInvite.getTo());
-            params.putString("call_state", activeCallInvite.getState().name());
-            clearIncomingNotification(activeCallInvite);
+            params.putString("call_state", "BUSY");
+            clearIncomingNotification(activeCallInvite.getCallSid());
         }
         eventManager.sendEvent(EVENT_CONNECTION_DID_DISCONNECT, params);
     }
@@ -673,7 +711,11 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
             }
         }
 
-        activeCall = Voice.call(getReactApplicationContext(), accessToken, twiMLParams, callListener);
+        ConnectOptions connectOptions = new ConnectOptions.Builder(accessToken)
+            .params(twiMLParams)
+            .build();
+
+        activeCall = Voice.connect(getReactApplicationContext(), connectOptions, callListener);
     }
 
     @ReactMethod
@@ -705,22 +747,30 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                 Log.d(TAG, "Active call found state = "+activeCall.getState());
             }
             WritableMap params = Arguments.createMap();
+            String toNum = activeCall.getTo();
+            if (toNum == null) {
+                toNum = toNumber;
+            }
             params.putString("call_sid",   activeCall.getSid());
             params.putString("call_from",  activeCall.getFrom());
-            params.putString("call_to",    activeCall.getTo());
+            params.putString("call_to",    toNum);
             params.putString("call_state", activeCall.getState().name());
             promise.resolve(params);
             return;
         }
+        promise.resolve(null);
+    }
+
+    @ReactMethod
+    public void getCallInvite(Promise promise) {
         if (activeCallInvite != null) {
             if (BuildConfig.DEBUG) {
-                Log.d(TAG, "Active call invite found state = "+activeCallInvite.getState());
+                Log.d(TAG, "Call invite found "+ activeCallInvite);
             }
             WritableMap params = Arguments.createMap();
             params.putString("call_sid",   activeCallInvite.getCallSid());
             params.putString("call_from",  activeCallInvite.getFrom());
             params.putString("call_to",    activeCallInvite.getTo());
-            params.putString("call_state", activeCallInvite.getState().name());
             promise.resolve(params);
             return;
         }
@@ -734,8 +784,17 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
         audioManager.setSpeakerphoneOn(value);
     }
 
+    @ReactMethod
+    public void setOnHold(Boolean value) {
+        if (activeCall != null) {
+            activeCall.hold(value);
+        }
+    }
+
     private void setAudioFocus() {
         if (audioManager == null) {
+            audioManager.setMode(originalAudioMode);
+            audioManager.abandonAudioFocus(null);
             return;
         }
         originalAudioMode = audioManager.getMode();
@@ -755,11 +814,12 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
                 .build();
             audioManager.requestAudioFocus(focusRequest);
         } else {
-            audioManager.requestAudioFocus(
-                null,
-                AudioManager.STREAM_VOICE_CALL,
-                AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_EXCLUSIVE
-            );
+            int focusRequestResult = audioManager.requestAudioFocus(new AudioManager.OnAudioFocusChangeListener() {
+                @Override
+                public void onAudioFocusChange(int focusChange) {}
+            },
+            AudioManager.STREAM_VOICE_CALL,
+            AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
         }
         /*
          * Start by setting MODE_IN_COMMUNICATION as default audio mode. It is
@@ -772,6 +832,8 @@ public class TwilioVoiceModule extends ReactContextBaseJavaModule implements Act
 
     private void unsetAudioFocus() {
         if (audioManager == null) {
+            audioManager.setMode(originalAudioMode);
+            audioManager.abandonAudioFocus(null);
             return;
         }
         audioManager.setMode(originalAudioMode);