+++ index.js
@@ -17,7 +17,8 @@ const _eventHandlers = {
     deviceDidReceiveIncoming: new Map(),
     connectionDidConnect: new Map(),
     connectionDidDisconnect: new Map(),
-    //iOS specific
+    callStateRinging: new Map(),
+    callInviteCancelled: new Map(),
     callRejected: new Map(),
 }
 
@@ -52,9 +53,7 @@ const Twilio = {
     connect(params = {}) {
         TwilioVoice.connect(params)
     },
-    disconnect() {
-        TwilioVoice.disconnect()
-    },
+    disconnect: TwilioVoice.disconnect,
     accept() {
         if (Platform.OS === IOS) {
             return
@@ -73,23 +72,17 @@ const Twilio = {
         }
         TwilioVoice.ignore()
     },
-    setMuted(isMuted) {
-        TwilioVoice.setMuted(isMuted)
-    },
-    setSpeakerPhone(value) {
-        TwilioVoice.setSpeakerPhone(value)
-    },
-    sendDigits(digits) {
-        TwilioVoice.sendDigits(digits)
-    },
+    setMuted: TwilioVoice.setMuted,
+    setSpeakerPhone: TwilioVoice.setSpeakerPhone,
+    sendDigits: TwilioVoice.sendDigits,
+    hold: TwilioVoice.hold,
     requestPermissions(senderId) {
         if (Platform.OS === ANDROID) {
             TwilioVoice.requestPermissions(senderId)
         }
     },
-    getActiveCall() {
-        return TwilioVoice.getActiveCall()
-    },
+    getActiveCall: TwilioVoice.getActiveCall,
+    getCallInvite: TwilioVoice.getCallInvite,
     configureCallKit(params = {}) {
         if (Platform.OS === IOS) {
             TwilioVoice.configureCallKit(params)
@@ -101,6 +94,10 @@ const Twilio = {
         }
     },
     addEventListener(type, handler) {
+        if (!_eventHandlers.hasOwnProperty(type)) {
+            throw new Error('Event handler not found: ' + type)
+        }
+        if (_eventHandlers[type])
         if (_eventHandlers[type].has(handler)) {
             return
         }